moved classes in job.py to seperate files, strategy.py, sync.py and util.py
